Connection: move connection custom caps to package
Move the connection-related custom capabilities to the Connection package.
This will allow the custom capabilities to be used when the package is required by
another plugin and Jetpack isn't installed.

Add unit tests for the new Manager::jetpack_connection_custom_caps method.

Add a new parameter, $namespace, to the ManagerTest::mock_function method. This allows
mocking of functions with different namespaces.